Refactor: extract TUF root client routines into a separate file
Initially, I planned to extract all TUF related routines.
But the TUF targets are so intervowen into everthing, that it comprises one third of the client.
That is still too much for one file, so targets deserve one more, maybe more splits.

The TUF root is still rather neat, and I plan quite some changes to these routines.
So, they are best off right now.
